以模版的形式储存html而不是大ftring
